Rename src/types/reducers to src/types/state
