+++ src/components/calltree/CallTree.js
@@ -28,7 +28,7 @@ import {
   addTransformToStack,
 } from '../../actions/profile-view';
 
-import type { IconWithClassName, State } from '../../types/reducers';
+import type { IconWithClassName, State } from '../../types/state';
 import type { CallTree } from '../../profile-logic/call-tree';
 import type { ImplementationFilter } from '../../types/actions';
 import type { ThreadIndex } from '../../types/profile';
