+++ src/components/timeline/TrackThread.js
@@ -44,7 +44,7 @@ import type {
   CallNodeInfo,
   IndexIntoCallNodeTable,
 } from '../../types/profile-derived';
-import type { State } from '../../types/reducers';
+import type { State } from '../../types/state';
 import type {
   ExplicitConnectOptions,
   ConnectedProps,
